feat(global-state): added global state managers for cart products and filtered search
The product store handles all the logic to update the products that are filtered dynamically in the
searchbar. On the other hand, the cart store is more complex, since it handles all the logic
required to add and delete items from the cart based on user selection. This global states is being
persisted in local storage so when the user refreshes the website, the selected items will remain in
the cart